+++ src/components/Muestras.vue
@@ -1351,6 +1351,16 @@ div
 
   Separador
 
+  #actividad-dialogo.titulo-segundo.color-acento-botones
+    h2 Actividad dialogo Ingles
+
+  Dialogo.color-primario(:dialogo="dialogo")
+    template(#nombre-actividad) Arrastrar y Soltar
+    template(#titulo-actividad) Titulo de actividad - Completa la conversación.
+    template(#descripcion-actividad) Arrastra y suelta el cada una de las palabras en el orden correcto.
+
+  Separador
+
   #lineas_tiempo.titulo-segundo.color-acento-botones
     h2 Líneas de tiempo
 
@@ -2337,6 +2347,56 @@ export default {
     indicadorTarjetaSlide: true,
     modal1: false,
     modal2: false,
+    dialogo: {
+      personajes: [
+        {
+          nombre: 'Hanna',
+          img: require('@/assets/componentes/ej-05.svg'),
+        },
+        {
+          nombre: 'Jhon',
+          img: require('@/assets/componentes/ej-05.svg'),
+        },
+      ],
+      dialogo: [
+        {
+          personaje: 'Hanna',
+          textoIng: 'I need the biology book and the sheets.',
+          textoEsp: 'Yo necesito el libro de biologia y las diapositivas.',
+          audio: require('@/assets/componentes/audios/audio-ej.mp3'),
+        },
+        {
+          personaje: 'Jhon',
+          textoIng:
+            'I need the *** book and the sheets book and the sheets book and the sheets book and the sheets.',
+          textoEsp:
+            'Yo necesito el libro de *** y las diapositivas Yo necesito el libro de Yo necesito el libro de Yo necesito el libro de .',
+          audio: require('@/assets/componentes/audios/audio-ej.mp3'),
+          palabra: 'Biology1',
+        },
+        {
+          personaje: 'Hanna',
+          textoIng: 'I need the biology books and the ***.',
+          textoEsp: 'Yo necesito el libro de biologia y las ***.',
+          audio: require('@/assets/componentes/audios/audio-ej.mp3'),
+          palabra: 'sheets2',
+        },
+        {
+          personaje: 'Jhon',
+          textoIng: 'I *** the biology books and the sheets',
+          textoEsp: 'Yo *** el libro de biologia y las diapositivas.',
+          audio: require('@/assets/componentes/audios/audio-ej.mp3'),
+          palabra: 'need3',
+        },
+        {
+          personaje: 'Hanna',
+          textoIng: 'I need the biology books and the ***.',
+          textoEsp: 'Yo necesito el libro de biologia y las ***.',
+          audio: require('@/assets/componentes/audios/audio-ej.mp3'),
+          palabra: 'sheets4',
+        },
+      ],
+    },
     datosLineaTiempoA: [
       {
         ano: '100.000 - 10.000 A.C.',